develop -> master

+++ everytime/article/views.py
@@ -1,20 +1,77 @@
 from re import U
 from django.core.exceptions import ObjectDoesNotExist
 from django.shortcuts import render
+from django.db.models import Q
+
 from rest_framework import serializers, status, viewsets, permissions
 from rest_framework.decorators import action, permission_classes
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from operator import itemgetter, attrgetter, methodcaller
 from django.db.models.query import QuerySet
-
+from datetime import datetime, timedelta
 from .serializers import *
 from university.models import University
 from board.models import Board
 #from .models import Article
 import requests
 
+import logging
+logger = logging.getLogger('django')
+
 # Create your views here.
+
+class ArticleWholeViewSet(viewsets.GenericViewSet):
+    serializer_class = ArticleCreateSerializer
+    permission_classes = (permissions.IsAuthenticated,)
+
+    def list(self, request):
+        queryset = self.get_queryset().order_by('-id')
+
+        # # 검색 기능
+        # search = self.request.query_params.get('search', None)
+        # queryset = self.get_queryset_search(search, queryset)
+
+        # # Hot 게시물, Best 게시물
+        # interest = self.request.query_params.get('interest', None)
+        # queryset = self.get_queryset_interest(interest, queryset)
+
+        search = request.query_params.get('search', None)
+
+        if not search or search == "":
+            return Response(status=status.HTTP_400_BAD_REQUEST, data="put search keyword")
+
+        # 검색 기능
+        if search:
+            queryset = self.get_queryset_search(search, queryset)
+        
+        return self.paginate(request, queryset)
+
+    def paginate(self, request, queryset):
+        page = self.paginate_queryset(queryset)
+
+        if page is not None:
+            serializer = ArticleSerializer(page, many=True, context={'request': request})
+            return self.get_paginated_response(serializer.data) 
+        else:
+            return Response(status=status.HTTP_400_BAD_REQUEST, data="pagination fault")
+    
+    def get_queryset(self):
+        queryset = Article.objects.all()
+        return queryset
+
+    def get_queryset_search(self, search, queryset):
+        if search == "":
+            return []
+        if search:
+            q = Q()
+            #return queryset.filter(q)
+            keywords = set(search.split(' '))
+            for k in keywords:
+               q &= Q(title__icontains=k)|Q(text__icontains=k)
+            queryset = queryset.filter(q)
+        return queryset.distinct()
+   
 class ArticleViewSet(viewsets.GenericViewSet):
     serializer_class = ArticleCreateSerializer
     permission_classes = (permissions.IsAuthenticated,)
@@ -29,8 +86,10 @@ class ArticleViewSet(viewsets.GenericViewSet):
         data['board'] = board_id
         serializer = self.get_serializer(data=data)
         serializer.is_valid(raise_exception=True)
-
         article = serializer.save()
+
+        nickname_code = 0 if article.is_anonymous == True else None
+        user_article = UserArticle.objects.create(article=article, user=request.user, subscribe=True, nickname_code=nickname_code)
 
         return Response(status=status.HTTP_200_OK, data={"success" : True, "article_id" : article.id})
 
@@ -63,6 +122,10 @@ class ArticleViewSet(viewsets.GenericViewSet):
 
         if article.board != board:
             return Response(status=status.HTTP_404_NOT_FOUND, data={ "error":"wrong_match", "detail" : "해당 게시판의 게시글이 아닙니다."})
+
+        if article.is_question and Comment.objects.filter(Q(article=article) & ~Q(article=article, commenter=request.user)):
+            return Response(status=status.HTTP_400_BAD_REQUEST, data={"error": "wrong_match", "detail": "댓글이 달린 질문글은 삭제가 불가능합니다. "})
+        
         article.delete()
 
         return Response(status=status.HTTP_200_OK, data={"success" : True})
@@ -74,14 +137,26 @@ class ArticleViewSet(viewsets.GenericViewSet):
 
         queryset = self.get_queryset().filter(board_id=board_id).order_by('-id')
 
-        # 검색 기능
+        # # 검색 기능
+        # search = self.request.query_params.get('search', None)
+        # queryset = self.get_queryset_search(search, queryset)
+
+        # # Hot 게시물, Best 게시물
+        # interest = self.request.query_params.get('interest', None)
+        # queryset = self.get_queryset_interest(interest, queryset)
+
         search = self.request.query_params.get('search', None)
-        queryset = self.get_queryset_search(search, queryset)
-
-        # Hot 게시물, Best 게시물
         interest = self.request.query_params.get('interest', None)
-        queryset = self.get_queryset_interest(interest, queryset)
 
+        # 검색 기능
+        if search:
+            queryset = self.get_queryset_search(search, queryset)
+        # Hot 게시물, Best 게시물
+        elif interest:
+            queryset = self.get_queryset_interest(interest, queryset)
+        elif interest and search:
+            return Response(status=status.HTTP_400_BAD_REQUEST, data={"detail" : "wrong query parameter"})
+        
         return self.paginate(request, queryset)
 
     def paginate(self, request, queryset):
@@ -101,16 +176,43 @@ class ArticleViewSet(viewsets.GenericViewSet):
         if search == "":
             return []
         if search:
-            queryset_, queryset = queryset, []
-            search = set(search.split(' '))
-            for article in queryset_:
-                article_words = set((article.title+" "+article.text).split(' '))
-                if search.issubset(article_words):
-                    queryset.append(article)
-        return queryset
+            q = Q()
+            #return queryset.filter(q)
+            keywords = set(search.split(' '))
+            for k in keywords:
+               q &= Q(title__icontains=k)|Q(text__icontains=k)
+            queryset = queryset.filter(q)
+        return queryset.distinct()
+            
+        # if search:
+        #     queryset_, queryset = queryset, []
+        #     search = set(search.split(' '))
+        #     for article in queryset_:
+        #         article_words = set((article.title+" "+article.text).split(' '))
+        #         if search.issubset(article_words):
+        #             queryset.append(article)
+        # return queryset
 
     def get_queryset_interest(self, interest, queryset):
-        if interest:
+        if interest == 'question':
+            queryset_, queryset = queryset, []
+            for article in queryset_:
+                if article.is_question:
+                    queryset.append(article)
+        if interest == 'live':
+            queryset_, queryset = queryset, []
+            for article in queryset_:
+                delta = (datetime.now() - article.created_at.replace(tzinfo=None))
+                if delta <= timedelta(days=1):
+                    queryset.append((article,
+                                     UserArticle.objects.filter(article=article, like=True).count(),
+                                     Comment.objects.filter(article=article).count()))
+
+            queryset_ = sorted(queryset, key=lambda x: x[1]+x[2], reverse=True)
+            queryset.clear()
+            for query in queryset_:
+                queryset.append(query[0])
+        elif interest:
             queryset_, queryset = queryset, []
             required_likes = {'hot': 5, 'best': 50}
             if interest in required_likes:
@@ -133,12 +235,24 @@ class ArticleViewSet_My_All(ArticleViewSet):
                             data={"error": "wrong_board_id", "detail": "게시판이 존재하지 않습니다."})
 
         # 검색 기능
-        search = self.request.query_params.get('search', None)
-        queryset = self.get_queryset_search(search, queryset)
+        #search = self.request.query_params.get('search', None)
+        #queryset = self.get_queryset_search(search, queryset)
 
         # Hot 게시물, Best 게시물
+        #interest = self.request.query_params.get('interest', None)
+        #queryset = self.get_queryset_interest(interest, queryset)
+
+        search = self.request.query_params.get('search', None)
         interest = self.request.query_params.get('interest', None)
-        queryset = self.get_queryset_interest(interest, queryset)
+
+        # 검색 기능
+        if search:
+            queryset = self.get_queryset_search(search, queryset)
+        # Hot 게시물, Best 게시물
+        elif interest:
+            queryset = self.get_queryset_interest(interest, queryset)
+        elif interest and search:
+            return Response(status=status.HTTP_400_BAD_REQUEST, data={"detail": "wrong query parameter"})
 
         return self.paginate(request, queryset)
 
@@ -189,7 +303,7 @@ class UserArticleView(viewsets.GenericViewSet):
             serializer = self.get_serializer(user_article, data={}, partial=True)
             serializer.is_valid(raise_exception=True)
             serializer.update(user_article, serializer.validated_data)
-        return self.get_response(article, user_article.scrap)
+        return self.get_response(article, user_article)
 
     def get_response(self, article):
         return Response(status=status.HTTP_200_OK)
@@ -197,7 +311,7 @@ class UserArticleView(viewsets.GenericViewSet):
 
 
 class UserArticleLikeView(UserArticleView):
-    def get_response(self, article, scrap):
+    def get_response(self, article, user_article):
         return Response(status=status.HTTP_200_OK,
                         data={
                             "like": UserArticle.objects.filter(article=article, like=True).count(),
@@ -207,10 +321,19 @@ class UserArticleLikeView(UserArticleView):
 
 
 class UserArticleScrapView(UserArticleView):
-    def get_response(self, article, scrap):
+    def get_response(self, article, user_article):
         return Response(status=status.HTTP_200_OK,
                         data={
                             "scrap": UserArticle.objects.filter(article=article, scrap=True).count(),
-                            "detail": "이 글을 스크랩하였습니다." if scrap else "스크랩을 취소하였습니다."
+                            "detail": "이 글을 스크랩하였습니다." if user_article.scrap else "스크랩을 취소하였습니다."
+                            }
+                        )
+
+class UserArticleSubscribeView(UserArticleView):
+    def get_response(self, article, user_article):
+        return Response(status=status.HTTP_200_OK,
+                        data={
+                            "subscribe": user_article.subscribe,
+                            "detail": "댓글 알림을 켰습니다." if user_article.subscribe else "댓글 알림을 껐습니다."
                             }
                         )
